requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
requests
requests-cache
os